# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Joke App API",
    "description": "Auto-generated Swagger UI for Joke App",
    "version": "1.0.0"
  },
  "host": "192.168.100.177:8010",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "description": "Type id"
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category name"
          },
          {
            "name": "color",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category color"
          },
          {
            "name": "border_color",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category border color"
          },
          {
            "name": "image_file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Category image"
          },
          {
            "name": "has_subcategory",
            "in": "formData",
            "type": "boolean",
            "required": true,
            "description": ""
          },
          {
            "name": "version",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category version"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type_id",
            "in": "formData",
            "type": "number",
            "required": true,
            "description": "Type id"
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category name"
          },
          {
            "name": "color",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category color"
          },
          {
            "name": "border_color",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category border color"
          },
          {
            "name": "image_file",
            "in": "formData",
            "type": "file",
            "required": false,
            "description": "Category image"
          },
          {
            "name": "cover_image",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category image"
          },
          {
            "name": "has_subcategory",
            "in": "formData",
            "type": "boolean",
            "required": true,
            "description": ""
          },
          {
            "name": "version",
            "in": "formData",
            "type": "string",
            "required": true,
            "description": "Category version"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cover_image": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}